requirements.txt
pymongo~=4.10.1
gradio~=4.44.1
pandas~=2.2.3
plotly~=5.24.1
ashe~=0.0.8
ruamel.yaml~=0.18.6
clickhouse_driver~=0.2.9